Migrate to Tyrian v2 based on Boostrap 4
Tyrian v2 is based on Boostrap 4, which is a major rewrite
of the whole project. For more information please have a
look at: https://getbootstrap.com/docs/4.0/migration/

Visually speaking Tyrian v2 is similar to Tyrian v1.

Speaking of this site (www.gentoo.de) the most prominent
change is the fixed sticky footer which is correctly
displayed in Tyrian v2. Apart from that there are only
minor changes in the visual appearance.

Signed-off-by: Max Magorsch <arzano@gentoo.org>
Closes: https://github.com/gentoo-ev/www.gentoo.de/pull/10
Signed-off-by: Thomas Deutschmann <whissi@gentoo.org>